Mise à jour du projet: connexion google

+++ database/migrations/2024_07_25_110906_create_messages_table.php
@@ -12,19 +12,18 @@ class CreateMessagesTable extends Migration
     public function up()
     {
         Schema::create('messages', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('place_id');
+            $table->bigIncrements('id'); // Utilisez bigIncrements si nécessaire
+            $table->unsignedBigInteger('user_id'); // Assurez-vous que c'est bigInteger
+            $table->unsignedBigInteger('place_id'); // Assurez-vous que c'est bigInteger
             $table->text('content');
             $table->string('status');
             $table->boolean('is_favorite')->default(false);
             $table->boolean('is_report')->default(false);
-            $table->timestamps(); // Ajoute les colonnes created_at et updated_at
+            $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('place_id')->references('id')->on('places')->onDelete('cascade');
         });
     }
-
     /**
      * Reverse the migrations.
      */
